# be/cmd/swag/docs/swagger.yaml
basePath: /
definitions:
...
  model.AddMemberToChatInput:
    properties:
      admin_chat_id:
        type: string
      chat_id:
        type: string
      user_add_id:
        type: string
    type: object
  model.ChangeAdminGroupChatInput:
    properties:
      group_chat_id:
        type: string
      new_admin_id:
        type: string
      old_admin_id:
        type: string
    type: object
  model.CreateChatGroupInput:
    properties:
      group_name:
        type: string
      list_id:
        items:
          type: string
        type: array
      user_id_create:
        type: string
    type: object
  model.CreateChatPrivateInput:
    properties:
      user1:
        type: string
      user2:
        type: string
    type: object
...
  model.DelChatInput:
    properties:
      admin_chat_id:
        type: string
      chat_id:
        type: string
    type: object
  model.DelMenForChatInput:
    properties:
      admin_chat_id:
        type: string
      chat_id:
        type: string
      user_del_id:
        type: string
    type: object
...
  model.UpgradeChatInfoInput:
    properties:
      group_avatar:
        type: string
      group_chat_id:
        type: string
      group_name_update:
        type: string
      user_admin_id:
        type: string
    type: object
...
  response.ErrResponseData:
    properties:
      code:
        description: Ma status code
        type: integer
      detail:
        description: Thong bao loi
      error:
        type: string
    type: object
  response.ResponseData:
    properties:
      code:
        description: Ma status code
        type: integer
      data:
        description: Du lieu duoc return
      message:
        description: Thong bao loi
        type: string
    type: object
...
host: localhost:8082
info:
  contact:
    email: lytranvinh.work@gmail.com
    name: TEAM V
    url: example.com/be/
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: example.com/be/
  title: API Documentation Chat App
  version: 1.0.0
paths:
  /api/v1/chat/add-member-to-chat:
    post:
      consumes:
      - application/json
      description: Add member to chat by admin group member
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.AddMemberToChatInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Add member to chat
      tags:
      - Chat
  /api/v1/chat/change-admin-group-chat:
    put:
      consumes:
      - application/json
      description: Change admin chat status from old admin to new admin
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.ChangeAdminGroupChatInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Change admin chat status
      tags:
      - Chat
  /api/v1/chat/create-chat-group:
    post:
      consumes:
      - application/json
      description: Create a chat group with user after auth, user create is admin
        and others users are members
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CreateChatGroupInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Create a chat group with user after auth
      tags:
      - Chat
  /api/v1/chat/create-chat-private:
    post:
      consumes:
      - application/json
      description: Create a chat private both user
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CreateChatPrivateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Create a chat private
      tags:
      - Chat
  /api/v1/chat/del-chat:
    delete:
      consumes:
      - application/json
      description: Delete chat by admin group chat
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.DelChatInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Delete chat
      tags:
      - Chat
  /api/v1/chat/del-men-from-chat:
    delete:
      consumes:
      - application/json
      description: Delete member from chat by admin group chat
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.DelMenForChatInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Delete member from chat
      tags:
      - Chat
  /api/v1/chat/get-chat-info:
    get:
      consumes:
      - application/json
      description: Get chat information by chat id
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Chat ID
        in: query
        name: chat_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Get chat information
      tags:
      - Chat
  /api/v1/chat/get-list-chat-for-user:
    get:
      consumes:
      - application/json
      description: Get list chat from user
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: limit number of chat
        in: query
        name: limit
        required: true
        type: string
      - description: page number
        in: query
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Get list of chat
      tags:
      - Chat
  /api/v1/chat/get-user-in-chat:
    get:
      consumes:
      - application/json
      description: Get information user in chat with chat id
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Chat ID
        in: query
        name: chat_id
        required: true
        type: string
      - description: limit number of chat
        in: query
        name: limit
        required: true
        type: string
      - description: page number
        in: query
        name: page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Hanlde get infomation user in chat
      tags:
      - Chat
  /api/v1/chat/upgrade-chat-info:
    put:
      consumes:
      - application/json
      description: Update chat info by admin group chat
      parameters:
      - description: Authorization Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpgradeChatInfoInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrResponseData'
      summary: Update chat info
      tags:
      - Chat
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"